refractor to monorepo pattern
