+++ src/essence/Tools/AddTempLayer/lib/index.ts
@@ -1,0 +1,9 @@
+// Components
+export { AddTempLayerPanel, type AddTempLayerPanelProps } from './geo/AddTempLayerPanel/AddTempLayerPanel'
+export { AddLayerModal, type AddLayerModalProps } from './geo/AddLayerModal/AddLayerModal'
+
+// Shared domain types
+export type { TempLayerType, AddTempLayerInput } from './types'
+
+// Side-effect import of compiled styles
+import './styles/index.scss'
