+++ src/essence/Tools/AddTempLayer/lib/styles/components-geo/add-temp-layer.scss
@@ -1,0 +1,41 @@
+// Trigger button (dashed, full-width) that opens the add-layer modal.
+
+.blocks-add-temp-layer {
+    box-sizing: border-box;
+    width: 100%;
+    padding: var(--theme-spacing-1, 8px);
+    font-family: var(--theme-font-ui);
+
+    &__trigger {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        gap: var(--theme-spacing-1, 8px);
+        box-sizing: border-box;
+        width: 100%;
+        padding: var(--theme-spacing-1, 8px) var(--theme-spacing-2, 16px);
+        background: transparent;
+        border: 1px dashed var(--theme-color-base-light, #a9aeb1);
+        border-radius: var(--theme-radius-md, 4px);
+        color: var(--theme-color-primary, #137480);
+        font-family: var(--theme-font-ui);
+        font-size: var(--theme-font-size-2xs, 14px);
+        font-weight: var(--theme-font-weight-semibold, 600);
+        cursor: pointer;
+        transition: border-color 0.12s ease, background 0.12s ease, color 0.12s ease;
+
+        &:hover {
+            border-color: var(--theme-color-primary-dark, #0b4c54);
+            color: var(--theme-color-primary-dark, #0b4c54);
+            background: var(--theme-color-base-lightest, #f6f7f8);
+        }
+
+        svg {
+            flex-shrink: 0;
+        }
+    }
+
+    &__trigger-label {
+        white-space: nowrap;
+    }
+}
