+++ src/essence/Tools/AddTempLayer/lib/styles/components-geo/index.scss
@@ -1,0 +1,3 @@
+// Aggregator — one @use per component partial in this directory.
+@use 'add-temp-layer';
+@use 'add-layer-modal';
