+++ src/essence/Tools/AddTempLayer/lib/styles/index.scss
@@ -1,0 +1,4 @@
+// Theming (USWDS framework + theme tokens + :root --theme-* custom properties)
+// is provided by MMGIS's per-theme bundles at dist/<theme>.css, loaded at
+// runtime. Component partials reference --theme-* directly.
+@forward 'components-geo';
