+++ src/essence/Tools/AddTempLayer/lib/styles/scss-imports.d.ts
@@ -1,0 +1,1 @@
+declare module '*.scss'
