+++ src/essence/Tools/AddTempLayer/lib/types.ts
@@ -1,0 +1,8 @@
+export type TempLayerType = 'xyz' | 'wms' | 'wmts' | 'geojson'
+
+/** What the modal collects from the user. */
+export interface AddTempLayerInput {
+    url: string
+    displayName?: string
+    type?: TempLayerType
+}
