+++ src/essence/Tools/AddTempLayer/lib/utils/icons.tsx
@@ -1,0 +1,41 @@
+// Inline Material/USWDS glyphs — portable, no sprite/MDI-font coupling.
+import React from 'react'
+
+export function Icon({ size = 18, children }: { size?: number; children: React.ReactNode }) {
+    return (
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width={size}
+            height={size}
+            viewBox="0 0 24 24"
+            fill="currentColor"
+            aria-hidden="true"
+        >
+            {children}
+        </svg>
+    )
+}
+
+export function PlusIcon() {
+    return (
+        <Icon size={16}>
+            <path d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z" />
+        </Icon>
+    )
+}
+
+export function LinkIcon() {
+    return (
+        <Icon>
+            <path d="M3.9 12c0-1.71 1.39-3.1 3.1-3.1h4V7H7c-2.76 0-5 2.24-5 5s2.24 5 5 5h4v-1.9H7c-1.71 0-3.1-1.39-3.1-3.1zM8 13h8v-2H8v2zm9-6h-4v1.9h4c1.71 0 3.1 1.39 3.1 3.1s-1.39 3.1-3.1 3.1h-4V17h4c2.76 0 5-2.24 5-5s-2.24-5-5-5z" />
+        </Icon>
+    )
+}
+
+export function CloseIcon() {
+    return (
+        <Icon>
+            <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z" />
+        </Icon>
+    )
+}
